fix(COV1): clippy, and an assertion the compiler could fold away
Five findings, all this wave's own: a redundant `u32` cast at the terrain mint
site, an unnecessary `mut`, two `clone`s on a `Copy` array, and -- the one worth
a sentence -- `assertions_on_constants` on "an officer hides for longer than it
leans out", which compared two literals. It is asserted over the duty CYCLE now,
600 samples of one period, which is the same claim about the thing that runs.

Whole workspace, all targets, forced fresh with an extra `--cfg`: **0 errors,
0 warnings** over **569** re-checked units in 2 m 40 s.

+++ crates/inf-physics/src/d3/ecs.rs
@@ -1587,7 +1587,7 @@ impl PhysicsBridge3D {
                     super::label::ColliderLabel::part(
                         super::label::ColliderFamily::Terrain,
                         guid,
-                        described.saturating_sub(1) as u32,
+                        described.saturating_sub(1),
                     ),
                 );
                 let Some(collider) = terrain_tile_collider(tile, res, span) else {
